[user] Move dashboard to separate component
